Split config.py into multiple files
