+++ tests/spellbook.py
@@ -31,7 +31,7 @@ from pathlib import Path
 from click.testing import CliRunner
 
 from packit.cli.packit_base import packit_base
-from packit.config import Config
+from packit.config.config import Config
 from packit.utils import cwd, run_command
 
 TESTS_DIR = Path(__file__).parent
