+++ tests_recording/testbase.py
@@ -9,8 +9,8 @@ from requre.storage import PersistentObjectStorage
 
 import packit.distgit
 import packit.upstream
-from packit.config import Config
-from packit.config import get_package_config_from_repo
+from packit.config.config import Config
+from packit.config.package_config import get_package_config_from_repo
 from packit.exceptions import PackitException
 from packit.local_project import LocalProject
